feat: создает модуль с ошибками
