+++ js/main.js
@@ -2,3 +2,4 @@ import './img-upload-form.js';
 import { getData } from './api.js';
 
 getData();
+
